+++ ChestsAnywhere/Properties/AssemblyInfo.cs
@@ -1,5 +1,4 @@
 using System.Reflection;
 
 [assembly: AssemblyTitle("ChestsAnywhere")]
-[assembly: AssemblyVersion("1.15.1")]
-[assembly: AssemblyFileVersion("1.15.1")]
+[assembly: AssemblyVersion("1.16.0")]