+++ CropsAnytimeAnywhere/Properties/AssemblyInfo.cs
@@ -2,4 +2,3 @@ using System.Reflection;
 
 [assembly: AssemblyTitle("CropsAnytimeAnywhere")]
 [assembly: AssemblyVersion("1.1.0")]
-[assembly: AssemblyFileVersion("1.1.0")]