+++ DebugMode/Properties/AssemblyInfo.cs
@@ -1,5 +1,4 @@
 using System.Reflection;
 
 [assembly: AssemblyTitle("DebugMode")]
-[assembly: AssemblyVersion("1.9.3")]
-[assembly: AssemblyFileVersion("1.9.3")]
+[assembly: AssemblyVersion("1.9.4")]