+++ LookupAnything/Properties/AssemblyInfo.cs
@@ -1,5 +1,4 @@
 using System.Reflection;
 
 [assembly: AssemblyTitle("LookupAnything")]
-[assembly: AssemblyVersion("1.21.1")]
-[assembly: AssemblyFileVersion("1.21.1")]
+[assembly: AssemblyVersion("1.21.2")]