+++ RotateToolbar/Properties/AssemblyInfo.cs
@@ -2,4 +2,3 @@ using System.Reflection;
 
 [assembly: AssemblyTitle("RotateToolbar")]
 [assembly: AssemblyVersion("1.3.2")]
-[assembly: AssemblyFileVersion("1.3.2")]