+++ SkipIntro/Properties/AssemblyInfo.cs
@@ -2,4 +2,3 @@ using System.Reflection;
 
 [assembly: AssemblyTitle("SkipIntro")]
 [assembly: AssemblyVersion("1.8.2")]
-[assembly: AssemblyFileVersion("1.8.2")]