+++ SmallBeachFarm/Properties/AssemblyInfo.cs
@@ -2,4 +2,3 @@ using System.Reflection;
 
 [assembly: AssemblyTitle("SmallBeachFarm")]
 [assembly: AssemblyVersion("1.0.0")]
-[assembly: AssemblyFileVersion("1.0.0")]